backend/pyproject.toml
[tool.poetry]
name = "urlshortener"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
mysqlclient = "^2.2.6"
python-decouple = "^3.8"
django-allauth = {extras = ["socialaccount"], version = "^65.3.1"}
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"


[tool.poetry.group.dev.dependencies]
django-stubs = "^5.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
